[#21] 파일 및 함수 정리, 주석 추가
