+++ PomoTodo/PomoTodo/Sources/Views/Pomo/TimerSetupView.swift
@@ -12,13 +12,16 @@ struct TimerSetupView: View {
     
     var body: some View {
         VStack(alignment: .center) {
+            // 상단 태그
             PomoTopView()
             Spacer().frame(height: DynamicPadding.getHeight(36))
             
+            // 중앙 타이머
             PomoTimerView()
             Spacer().frame(height: DynamicPadding.getHeight(64))
             
+            // 버튼
             PomoBottomView()
-        }
+        } // : vstack
     }
 }
